+++ src/new.rs
@@ -7,7 +7,7 @@ use crate::Error;
 /// [`Record`]: trait.Record.html
 pub trait New<T>: Sized {
     /// Instantiate and returns a new `Self` instance from `T`.
-    /// Can fail and return an error, the error is in most of the cases a [`Error`]::[`ValidationError`]
+    /// Can fail and return an error, the error is in most of the cases an [`Error`]::[`ValidationError`]
     /// on fields validation failure
     ///
     /// [`Error`]: eum.Error.html